+++ core/domain/src/main/java/com/paradox543/malankaraorthodoxliturgica/domain/bible/model/PrefaceContent.kt
@@ -3,6 +3,6 @@ package com.paradox543.malankaraorthodoxliturgica.domain.bible.model
 import com.paradox543.malankaraorthodoxliturgica.domain.prayer.model.PrayerElement
 
 data class PrefaceContent(
-    val en: List<PrayerElement>,
-    val ml: List<PrayerElement>,
+    val en: List<PrayerElement.Prose>,
+    val ml: List<PrayerElement.Prose>,
 )